+++ R/find_pattern.R
@@ -13,6 +13,7 @@
 #' be interpreted as a \code{perl} regex? 
 #' @param file_contents_fixed (logical, default: \code{FALSE}) Should \code{file_contents} 
 #' be interpreted as a \code{fixed} regex?
+#' @param file_contents_ignore_case (logical, default: \code{FALSE}) As in \code{\link[base]{grep}}.
 #' @param file.ext A file extension passed to the operating system if \code{use.OS} is used.
 #' @param which_lines One of \code{"first"} and \code{"all"}. If \code{"first"} only the first match in any file is returned in the result; if \code{"all"}, all matches are.
 #' @return A \code{data.table}, showing the matches per file.
@@ -28,8 +29,15 @@ find_pattern_in <- function(file_contents,
                             file_pattern = "\\.(R|r)(nw|md)?$",
                             file_contents_perl = TRUE,
                             file_contents_fixed = FALSE,
+                            file_contents_ignore_case = FALSE,
                             file.ext = NULL, 
                             which_lines = c("first", "all")) {
+  # Harmonize perl,fixed,ignore_case
+  if (file_contents_fixed && missing(file_contents_perl)) {
+    file_contents_perl <- FALSE
+  }
+
+  
   ..reader <- match.fun(reader)
   .reader <- function(x) {
     out <- ..reader(x)
@@ -109,12 +117,14 @@ find_pattern_in <- function(file_contents,
              grep(.pattern,
                   x = x,
                   perl = file_contents_perl,
+                  ignore.case = file_contents_ignore_case,
                   fixed = file_contents_fixed)[1L]
            }, 
            "all" = {
              grep(.pattern,
                   x = x,
                   perl = file_contents_perl,
+                  ignore.case = file_contents_ignore_case,
                   fixed = file_contents_fixed)
            })
   }
@@ -126,6 +136,7 @@ find_pattern_in <- function(file_contents,
                   x = x,
                   value = TRUE,
                   perl = file_contents_perl,
+                  ignore.case = file_contents_ignore_case,
                   fixed = file_contents_fixed)[1L]
            }, 
            "all" = {
@@ -133,6 +144,7 @@ find_pattern_in <- function(file_contents,
                   x = x,
                   value = TRUE,
                   perl = file_contents_perl,
+                  ignore.case = file_contents_ignore_case,
                   fixed = file_contents_fixed)
            })
   }
@@ -167,3 +179,7 @@ deprecate_windows_finder <- function() {
                 "windows")) .Deprecated("hutilsInteractive::find_pattern_in_windows")
 }
 #nocov end
+
+
+
+